streamlit-cloudrun/pyproject.toml
[tool.poetry]
name = "streamlit-cloudrun"
version = "0.1.0"
description = ""
authors = ["ryutah <6662577+ryutah@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.35.0"
google-cloud-logging = "^3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
